S3 6 2

+++ backend/CulturalHeritage/src/main/java/tim2/CulturalHeritage/service/CHSubtypeServiceImpl.java
@@ -1,12 +1,16 @@
 package tim2.CulturalHeritage.service;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import tim2.CulturalHeritage.model.CHSubtype;
 import tim2.CulturalHeritage.repository.CHSubtypeRepository;
+
+import javax.persistence.EntityNotFoundException;
 
 @Service
 public class CHSubtypeServiceImpl implements CHSubtypeService {
@@ -25,7 +29,11 @@ public class CHSubtypeServiceImpl implements CHSubtypeService {
     }
 
     @Override
-    public CHSubtype add(CHSubtype chSubtype) {
+    public CHSubtype add(CHSubtype chSubtype) throws DataIntegrityViolationException {
+        CHSubtype existName = chSubtypeRepository.findByName(chSubtype.getName());
+        if(existName != null) {
+            throw new DataIntegrityViolationException("Name must be unique");
+        }
         return chSubtypeRepository.save(chSubtype);
     }
 
@@ -36,6 +44,11 @@ public class CHSubtypeServiceImpl implements CHSubtypeService {
 
     @Override
     public void deleteById(Long id) {
+        Optional<CHSubtype> exists = chSubtypeRepository.findById(id);
+        if(exists.isEmpty()) {
+            throw new EntityNotFoundException("There is no subtype with given id");
+        }
+
         chSubtypeRepository.deleteById(id);
     }
